refactor: move txn domain_metadata fns to separate file
